requirements.txt
python-jose>=3.3.0
aiohttp>=3.8.6
cryptography>=41.0.5
